move single-use funcs to the pull cmd
