+++ commands/command_checkout.go
@@ -38,52 +38,6 @@ func checkoutCommand(cmd *cobra.Command, args []string) {
 	checkoutWithIncludeExclude(filter)
 }
 
-func checkoutFromFetchChan(in chan *lfs.WrappedPointer, filter *filepathfilter.Filter) {
-	ref, err := git.CurrentRef()
-	if err != nil {
-		Panic(err, "Could not checkout")
-	}
-
-	// Need to ScanTree to identify multiple files with the same content (fetch will only report oids once)
-	// use new gitscanner so mapping has all the scanned pointers before continuing
-	mapping := make(map[string][]*lfs.WrappedPointer)
-	chgitscanner := lfs.NewGitScanner(func(p *lfs.WrappedPointer, err error) {
-		if err != nil {
-			Panic(err, "Could not scan for Git LFS files")
-			return
-		}
-		mapping[p.Oid] = append(mapping[p.Oid], p)
-	})
-	chgitscanner.Filter = filter
-
-	if err := chgitscanner.ScanTree(ref.Sha); err != nil {
-		ExitWithError(err)
-	}
-
-	chgitscanner.Close()
-
-	// Launch git update-index
-	c := make(chan *lfs.WrappedPointer)
-
-	var wait sync.WaitGroup
-	wait.Add(1)
-
-	go func() {
-		checkoutWithChan(c)
-		wait.Done()
-	}()
-
-	// Feed it from in, which comes from fetch
-	for p := range in {
-		// Add all of the files for this oid
-		for _, fp := range mapping[p.Oid] {
-			c <- fp
-		}
-	}
-	close(c)
-	wait.Wait()
-}
-
 func checkoutWithIncludeExclude(filter *filepathfilter.Filter) {
 	ref, err := git.CurrentRef()
 	if err != nil {
